chore: created Attached Photo Preview Modal
